feat: split ethereum tests by categories: gov, impls, periphery, roles and config
